Bugs solved

+++ app/src/main/java/com/gabriel/weather/Wind.kt
@@ -2,6 +2,6 @@ package com.gabriel.weather
 
 data class Wind(
     val deg: Int,
-    val gust: Int,
+    val gust: Float,
     val speed: Double
 )
